+++ src/Font/Frontend/File/Table/HHeaTable.php
@@ -23,11 +23,11 @@ namespace PdfGenerator\Font\Frontend\File\Table;
 class HHeaTable
 {
     /**
-     * number of tables contained.
+     * version of table.
      *
-     * @ttf-type uint16
+     * @ttf-type fixed
      *
-     * @var int
+     * @var float
      */
     private $version;
 
@@ -127,7 +127,7 @@ class HHeaTable
      *
      * @ttf-type fword
      *
-     * @var float
+     * @var int
      */
     private $caretOffset;
 
@@ -150,12 +150,12 @@ class HHeaTable
      */
     private $numOfLongHorMetrics;
 
-    public function getVersion(): int
+    public function getVersion(): float
     {
         return $this->version;
     }
 
-    public function setVersion(int $version): void
+    public function setVersion(float $version): void
     {
         $this->version = $version;
     }
@@ -250,12 +250,12 @@ class HHeaTable
         $this->caretSlopeRun = $caretSlopeRun;
     }
 
-    public function getCaretOffset(): float
+    public function getCaretOffset(): int
     {
         return $this->caretOffset;
     }
 
-    public function setCaretOffset(float $caretOffset): void
+    public function setCaretOffset(int $caretOffset): void
     {
         $this->caretOffset = $caretOffset;
     }
